requirements.txt
streamlit>=1.34.0
requests>=2.31.0
sentence-transformers>=3.0.0
faiss-cpu>=1.7.4
pandas>=2.2.0
transformers>=4.42.0
torch>=2.2.0
python-docx>=1.0.1
reportlab>=4.0.9
accelerate>=0.30.0
safetensors>=0.4.2
bitsandbytes>=0.43.0